renamed SSH to skyssh
