+++ README.md
@@ -54,7 +54,7 @@ $ make
 $ OPTS="GSO111MODULE=on GOOS=linux GOARCH=arm" make
 ```
 
-**Install skywire-node, skywire-cli, manager-node and SSH-cli**
+**Install skywire-node, skywire-cli, manager-node and skyssh-cli**
 
 ```bash
 $ make install  # compiles and installs all binaries
@@ -114,7 +114,7 @@ After `skywire-node` is up and running with default environment, default apps ar
 - [Chat](/cmd/apps/skychat)
 - [Hello World](/cmd/apps/helloworld)
 - [The Real Proxy](/cmd/apps/therealproxy) ([Client](/cmd/apps/therealproxy-client))
-- [The Real SSH](/cmd/apps/SSH-server) ([Client](/cmd/apps/SSH-client))
+- [The Real skyssh](/cmd/apps/skyssh-server) ([Client](/cmd/apps/skyssh-client))
 
 ### Transports
 
@@ -186,12 +186,12 @@ This will:
 │   ├── helloworld.v1.0             #
 │   ├── socksproxy-client.v1.0    #
 │   ├── socksproxy.v1.0           #
-│   ├── SSH-client.v1.0      #
-│   └── SSH-server.v1.0             #
+│   ├── skyssh-client.v1.0      #
+│   └── skyssh-server.v1.0             #
 ├── local                           # **Created inside docker**
 │   ├── skychat                        #  according to "local_path" in skywire-config.json
 │   ├── socksproxy                  #
-│   └── SSH-server                  #
+│   └── skyssh-server                  #
 ├── PK                              # contains public key of node
 ├── skywire                         # db & logs. **Created inside docker**
 │   ├── routing.db                  #
@@ -292,8 +292,8 @@ $ GO111MODULE=on GOOS=linux go build -o /tmp/SKYNODE/skywire-node ./cmd/skywire-
 $ GO111MODULE=on GOOS=linux go build -o /tmp/SKYNODE/apps/skychat.v1.0 ./cmd/apps/skychat
 $ GO111MODULE=on GOOS=linux go build -o /tmp/SKYNODE/apps/helloworld.v1.0 ./cmd/apps/helloworld
 $ GO111MODULE=on GOOS=linux go build -o /tmp/SKYNODE/apps/therealproxy.v1.0 ./cmd/apps/therealproxy
-$ GO111MODULE=on GOOS=linux go build -o /tmp/SKYNODE/apps/SSH-server.v1.0  ./SSH/cmd/server
-$ GO111MODULE=on GOOS=linux go build -o /tmp/SKYNODE/apps/SSH-client.v1.0  ./SSH/cmd/client
+$ GO111MODULE=on GOOS=linux go build -o /tmp/SKYNODE/apps/skyssh-server.v1.0  ./skyssh/cmd/server
+$ GO111MODULE=on GOOS=linux go build -o /tmp/SKYNODE/apps/skyssh-client.v1.0  ./skyssh/cmd/client
 # 4. Create skywire-config.json for node
 $ skywire-cli node gen-config -o /tmp/SKYNODE/skywire-config.json
 # 2019/03/15 16:43:49 Done!
@@ -303,8 +303,8 @@ $ tree /tmp/SKYNODE
 # │   ├── skychat.v1.0
 # │   ├── helloworld.v1.0
 # │   ├── socksproxy.v1.0
-# │   ├── SSH-client.v1.0
-# │   └── SSH-server.v1.0
+# │   ├── skyssh-client.v1.0
+# │   └── skyssh-server.v1.0
 # ├── skywire-config.json
 # └── skywire-node
 # So far so good. We prepared docker volume. Now we can:
@@ -315,15 +315,15 @@ $ docker run -it -v /tmp/SKYNODE:/sky --network=SKYNET --name=SKYNODE skywire-ru
 # [2019-03-15T13:55:10Z] INFO [skywire]: Starting skychat.v1.0
 # [2019-03-15T13:55:10Z] INFO [skywire]: Starting RPC interface on 127.0.0.1:3435
 # [2019-03-15T13:55:10Z] INFO [skywire]: Starting therealproxy.v1.0
-# [2019-03-15T13:55:10Z] INFO [skywire]: Starting SSH-server.v1.0
+# [2019-03-15T13:55:10Z] INFO [skywire]: Starting skyssh-server.v1.0
 # [2019-03-15T13:55:10Z] INFO [skywire]: Starting packet router
 # [2019-03-15T13:55:10Z] INFO [router]: Starting router
 # [2019-03-15T13:55:10Z] INFO [trmanager]: Starting transport manager
 # [2019-03-15T13:55:10Z] INFO [router]: Got new App request with type Init: {"app-name":"skychat",# "app-version":"1.0","protocol-version":"0.0.1"}
 # [2019-03-15T13:55:10Z] INFO [router]: Handshaked new connection with the app skychat.v1.0
 # [2019-03-15T13:55:10Z] INFO [skychat.v1.0]: 2019/03/15 13:55:10 Serving HTTP on :8000
-# [2019-03-15T13:55:10Z] INFO [router]: Got new App request with type Init: {"app-name":"SSH-server",# "app-version":"1.0","protocol-version":"0.0.1"}
-# [2019-03-15T13:55:10Z] INFO [router]: Handshaked new connection with the app SSH-server.v1.0
+# [2019-03-15T13:55:10Z] INFO [router]: Got new App request with type Init: {"app-name":"skyssh-server",# "app-version":"1.0","protocol-version":"0.0.1"}
+# [2019-03-15T13:55:10Z] INFO [router]: Handshaked new connection with the app skyssh-server.v1.0
 # [2019-03-15T13:55:10Z] INFO [router]: Got new App request with type Init: {"app-name":"socksproxy",# "app-version":"1.0","protocol-version":"0.0.1"}
 # [2019-03-15T13:55:10Z] INFO [router]: Handshaked new connection with the app socksproxy.v1.0
 ```
