--- SSH/cmd/SSH-cli/README.md
@@ -1,6 +1,0 @@
-# CLI for SSH app
-
-`SSH-cli` implements PTY related operations for SSH app.
-
-It connects to SSH client's RPC server, configures terminal for raw pty
-data and handles pty data exchange.
