--- SSH/cmd/SSH-cli/SSH-cli.go
@@ -1,12 +1,0 @@
-/*
-CLI for SSH app
-*/
-package main
-
-import (
-	"github.com/skycoin/skywire/SSH/cmd/SSH-cli/commands"
-)
-
-func main() {
-	commands.Execute()
-}
