--- SSH/cmd/SSH-client/README.md
@@ -1,8 +1,0 @@
-# Skywire SSH client app
-
-`SSH-client` app implements client for the SSH app. 
-
-It starts RCP interface for `SSH-cli` and handles incoming requests to
-the remote node via `skywire` connection.
-
-Please check docs for `SSH-server` app for further instructions.
