+++ skyssh/cmd/skyssh-cli/README.md
@@ -1,0 +1,6 @@
+# CLI for skyssh app
+
+`skyssh-cli` implements PTY related operations for skyssh app.
+
+It connects to skyssh client's RPC server, configures terminal for raw pty
+data and handles pty data exchange.
