+++ skyssh/cmd/skyssh-cli/skyssh-cli.go
@@ -1,0 +1,12 @@
+/*
+CLI for skyssh app
+*/
+package main
+
+import (
+	"github.com/skycoin/skywire/skyssh/cmd/skyssh-cli/commands"
+)
+
+func main() {
+	commands.Execute()
+}
