+++ skyssh/cmd/skyssh-client/README.md
@@ -1,0 +1,8 @@
+# Skywire skyssh client app
+
+`skyssh-client` app implements client for the skyssh app. 
+
+It starts RCP interface for `skyssh-cli` and handles incoming requests to
+the remote node via `skywire` connection.
+
+Please check docs for `skyssh-server` app for further instructions.
