+++ skyssh/cmd/skyssh-server/README.md
@@ -1,17 +1,17 @@
-# Skywire SSH app
+# Skywire skyssh app
 
-`SSH-server` app implements SSH functionality over skywirenet.
+`skyssh-server` app implements skyssh functionality over skywirenet.
 
-`SSH-cli` is used to initiate communication via client RPC
-exposed by `SSH` app. 
+`skyssh-cli` is used to initiate communication via client RPC
+exposed by `skyssh` app. 
 
-`SSH` app implements common SSH operations:
+`skyssh` app implements common skyssh operations:
 
 - starting remote shell
 - and executing commands remotely
 
 PubKey whitelisting is performed by adding public key to the
-authentication file (`$HOME/.SSH/authorized_keys` by default).
+authentication file (`$HOME/.skyssh/authorized_keys` by default).
 
 ** Local setup
 
@@ -22,7 +22,7 @@ Create 2 node config files:
 ```json
   "apps": [
     {
-      "app": "SSH-server",
+      "app": skyssh,
       "version": "1.0",
       "auto_start": true,
       "port": 2
@@ -35,7 +35,7 @@ Create 2 node config files:
 ```json
   "apps": [
     {
-      "app": "SSH-client",
+      "app": skyssh-client,
       "version": "1.0",
       "auto_start": true,
       "port": 22
@@ -46,9 +46,9 @@ Create 2 node config files:
 Compile binaries and start 2 nodes:
 
 ```bash
-$ go build -o apps/SSH-server.v1.0 ./SSH/cmd/SSH-server
-$ go build -o apps/SSH-client.v1.0 ./SSH/cmd/SSH-client
-$ go build ./SSH/cmd/SSH-cli
+$ go build -o apps/skyssh-server.v1.0 ./skyssh/cmd/skyssh-server
+$ go build -o apps/skyssh-client.v1.0 ./skyssh/cmd/skyssh-client
+$ go build ./skyssh/cmd/skyssh-cli
 $ ./skywire-node skywire1.json
 $ ./skywire-node skywire2.json
 ```
@@ -56,14 +56,14 @@ $ ./skywire-node skywire2.json
 Add public key of the second node to the auth file:
 
 ```bash
-$ mkdir ~/.SSH
-$ echo "0348c941c5015a05c455ff238af2e57fb8f914c399aab604e9abb5b32b91a4c1fe" > ~/.SSH/authorized_keys
+$ mkdir ~/.skyssh
+$ echo "0348c941c5015a05c455ff238af2e57fb8f914c399aab604e9abb5b32b91a4c1fe" > ~/.skyssh/authorized_keys
 ```
 
 Connect to the first node using CLI:
 
 ```bash
-$ ./SSH-cli 024ec47420176680816e0406250e7156465e4531f5b26057c9f6297bb0303558c7
+$ ./skyssh-cli 024ec47420176680816e0406250e7156465e4531f5b26057c9f6297bb0303558c7
 ```
 
 This should get you to the $HOME folder of the user(you in this case), which
