+++ skyssh/internal/therealssh/client.go
@@ -152,7 +152,7 @@ type RPCClient struct {
 
 // RequestPTY defines RPC request for a new PTY session.
 func (rpc *RPCClient) RequestPTY(args *RequestPTYArgs, channelID *uint32) error {
-	debug("requesting SSH channel")
+	debug("requesting skyssh channel")
 	localID, channel, err := rpc.c.OpenChannel(args.RemotePK)
 	if err != nil {
 		return err
