+++ skyssh/internal/therealssh/shell.go
