+++ skyssh/internal/therealssh/shell_darwin.go
